\name{sample1-package}
\alias{sample1-package}
\alias{sample1}
\docType{package}
\title{
\packageTitle{sample1}
}
\description{
\packageDescription{sample1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sample1}
\packageIndices{sample1}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{sample1}

Maintainer: \packageMaintainer{sample1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
