% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/sample1_function.R
\name{sample1_function}
\alias{sample1_function}
\title{A sample function for checking the package creation through RStudio}
\usage{
sample1_function(input = TRUE)
}
\arguments{
\item{input}{Is the input TRUE/FALSE. Defaults to TRUE}
}
\description{
This function checks for the argument passed and displays the result accorddingly
}
\examples{
sample1_function()
}
\keyword{sample1}

